genre displaying properly on card

+++ server/public/scripts/controllers/manage.controller.js
@@ -1,4 +1,4 @@
-movieApp.controller('ManageController', ['$http', function($http) {
+movieApp.controller('ManageController', ['$http', function ($http) {
     let vm = this;
     console.log('running ManageController');
     vm.genreToSend = {};
@@ -9,11 +9,11 @@ movieApp.controller('ManageController', ['$http', function($http) {
             method: 'POST',
             url: '/manage',
             data: vm.genreToSend
-        }).then(function(response) {
+        }).then(function (response) {
             console.log('genre post back from server with', response);
             vm.genreToSend = {};
             vm.getGenres();
-        }).catch(function(error) {
+        }).catch(function (error) {
             console.log('error posting new genre to server', error);
         });
     }
@@ -22,24 +22,31 @@ movieApp.controller('ManageController', ['$http', function($http) {
         $http({
             method: 'GET',
             url: '/manage'
-        }).then(function(response) {
+        }).then(function (response) {
             console.log('back from server with genres', response.data);
             vm.genreArray = response.data;
-        }).catch(function(error) {
-            console.log('error getting genres from server', error); 
+        }).catch(function (error) {
+            console.log('error getting genres from server', error);
         });
     }
 
-    vm.deleteGenre = function (genreID) {
-        $http({
-            method: 'DELETE',
-            url: `/manage/${genreID}`
-        }).then(function(response) {
-            console.log('genre deleted from server', response);
-            vm.getGenres();
-        }).catch(function(error) {
-            console.log('error deleting genre from server', error);
-        });
+    vm.deleteGenre = function (genre) {
+        if (genre.count > 0) {
+            alert(`You can't delete a genre that has existing movies connected.`);
+        }
+        else {
+            $http({
+                method: 'DELETE',
+                url: `/manage`,
+                params: genre
+            }).then(function (response) {
+                console.log('genre deleted from server', response);
+                vm.getGenres();
+            }).catch(function (error) {
+                console.log('error deleting genre from server', error);
+            });
+        }
+
     }
 
     vm.getGenres();
